# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface GeneratedEmail {
  /** The generated email address */
  email: string;
  /** Username part of the email */
  username: string;
  /** Domain part of the email */
  domain: string;
  /** When the email session expires */
  expiresAt: string;
  /** When the email was generated */
  createdAt: string;
}

export interface EmailMessageSummary {
  id: string;
  from: string;
  subject: string;
  /** Short text preview of the email body */
  preview: string;
  receivedAt: string;
  isRead: boolean;
  hasAttachments: boolean;
}

export interface InboxResponse {
  email: string;
  messages: EmailMessageSummary[];
  total: number;
  unreadCount: number;
}

export interface Attachment {
  filename: string;
  contentType: string;
  size: number;
}

export interface EmailMessage {
  id: string;
  email: string;
  from: string;
  to: string;
  subject: string;
  textBody?: string;
  htmlBody?: string;
  receivedAt: string;
  isRead: boolean;
  attachments: Attachment[];
}

export interface MarkReadRequest {
  id: string;
  email: string;
}

export interface DomainsResponse {
  domains: string[];
}

export interface EmailStats {
  email: string;
  totalMessages: number;
  readCount: number;
  unreadCount: number;
  expiresAt: string;
  isExpired: boolean;
}

export interface ExtendEmailRequest {
  email: string;
  /** How many minutes to extend (default 30) */
  extraMinutes?: number;
}

export interface ExtendEmailResponse {
  email: string;
  newExpiresAt: string;
  extended: boolean;
}

export interface SuccessResponse {
  success: boolean;
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface BlacklistEntry {
  id: number;
  pattern: string;
  createdAt: string;
}

export interface BlacklistResponse {
  blocked: BlacklistEntry[];
}

export interface AddBlacklistRequest {
  email: string;
  /** Sender email or domain to block (e.g. spam@evil.com or @evil.com) */
  pattern: string;
}

export type GenerateEmailParams = {
  /**
   * Optional custom username
   */
  username?: string;
  /**
   * Optional domain selection
   */
  domain?: string;
};

export type GetInboxParams = {
  /**
   * The temporary email address
   */
  email: string;
};

export type GetMessageParams = {
  /**
   * The message ID
   */
  id: string;
  /**
   * The email address owning this message
   */
  email: string;
};

export type ResetInboxParams = {
  /**
   * The email address to reset
   */
  email: string;
};

export type GetBlacklistParams = {
  email: string;
};

export type RemoveFromBlacklistParams = {
  email: string;
  pattern: string;
};

export type GetEmailStatsParams = {
  /**
   * The temporary email address
   */
  email: string;
};