# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API specification
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Generate a temporary email address
 */
export const GenerateEmailQueryParams = zod.object({
  username: zod.coerce.string().optional().describe("Optional custom username"),
  domain: zod.coerce.string().optional().describe("Optional domain selection"),
});

export const GenerateEmailResponse = zod.object({
  email: zod.string().describe("The generated email address"),
  username: zod.string().describe("Username part of the email"),
  domain: zod.string().describe("Domain part of the email"),
  expiresAt: zod.coerce.date().describe("When the email session expires"),
  createdAt: zod.coerce.date().describe("When the email was generated"),
});

/**
 * @summary Get inbox for a temporary email address
 */
export const GetInboxQueryParams = zod.object({
  email: zod.coerce.string().describe("The temporary email address"),
});

export const GetInboxResponse = zod.object({
  email: zod.string(),
  messages: zod.array(
    zod.object({
      id: zod.string(),
      from: zod.string(),
      subject: zod.string(),
      preview: zod.string().describe("Short text preview of the email body"),
      receivedAt: zod.coerce.date(),
      isRead: zod.boolean(),
      hasAttachments: zod.boolean(),
    }),
  ),
  total: zod.number(),
  unreadCount: zod.number(),
});

/**
 * @summary Get a specific email message
 */
export const GetMessageQueryParams = zod.object({
  id: zod.coerce.string().describe("The message ID"),
  email: zod.coerce.string().describe("The email address owning this message"),
});

export const GetMessageResponse = zod.object({
  id: zod.string(),
  email: zod.string(),
  from: zod.string(),
  to: zod.string(),
  subject: zod.string(),
  textBody: zod.string().optional(),
  htmlBody: zod.string().optional(),
  receivedAt: zod.coerce.date(),
  isRead: zod.boolean(),
  attachments: zod.array(
    zod.object({
      filename: zod.string(),
      contentType: zod.string(),
      size: zod.number(),
    }),
  ),
});

/**
 * @summary Mark a message as read
 */
export const MarkMessageReadBody = zod.object({
  id: zod.string(),
  email: zod.string(),
});

export const MarkMessageReadResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Delete all messages in an inbox
 */
export const ResetInboxQueryParams = zod.object({
  email: zod.coerce.string().describe("The email address to reset"),
});

export const ResetInboxResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Extend the TTL of a temporary email address
 */
export const ExtendEmailBody = zod.object({
  email: zod.string(),
  extraMinutes: zod
    .number()
    .optional()
    .describe("How many minutes to extend (default 30)"),
});

export const ExtendEmailResponse = zod.object({
  email: zod.string(),
  newExpiresAt: zod.coerce.date(),
  extended: zod.boolean(),
});

/**
 * @summary Get list of available email domains
 */
export const GetAvailableDomainsResponse = zod.object({
  domains: zod.array(zod.string()),
});

/**
 * @summary Get blocked senders list for an email address
 */
export const GetBlacklistQueryParams = zod.object({
  email: zod.coerce.string(),
});

export const GetBlacklistResponse = zod.object({
  blocked: zod.array(
    zod.object({
      id: zod.number(),
      pattern: zod.string(),
      createdAt: zod.coerce.date(),
    }),
  ),
});

/**
 * @summary Block a sender or domain
 */
export const AddToBlacklistBody = zod.object({
  email: zod.string(),
  pattern: zod
    .string()
    .describe(
      "Sender email or domain to block (e.g. spam@evil.com or @evil.com)",
    ),
});

export const AddToBlacklistResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Remove a sender from blacklist
 */
export const RemoveFromBlacklistQueryParams = zod.object({
  email: zod.coerce.string(),
  pattern: zod.coerce.string(),
});

export const RemoveFromBlacklistResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
});

/**
 * @summary Get stats for a temporary email address (total, read, unread, expiry)
 */
export const GetEmailStatsQueryParams = zod.object({
  email: zod.coerce.string().describe("The temporary email address"),
});

export const GetEmailStatsResponse = zod.object({
  email: zod.string(),
  totalMessages: zod.number(),
  readCount: zod.number(),
  unreadCount: zod.number(),
  expiresAt: zod.coerce.date(),
  isExpired: zod.boolean(),
});