# lib/api-zod/src/generated/types/addBlacklistRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AddBlacklistRequest {
  email: string;
  /** Sender email or domain to block (e.g. spam@evil.com or @evil.com) */
  pattern: string;
}