# lib/api-zod/src/generated/types/blacklistEntry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API specification
 * OpenAPI spec version: 0.1.0
 */

export interface BlacklistEntry {
  id: number;
  pattern: string;
  createdAt: Date;
}