# lib/api-zod/src/generated/types/blacklistResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * TempMail API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BlacklistEntry } from "./blacklistEntry";

export interface BlacklistResponse {
  blocked: BlacklistEntry[];
}